Implement volatile change notifications
When a volatile signal is upgraded to a tracked source through
`subscribe` it gets an `onChange` handler. Until this commit it was
a no-op. Now the `onChange` handler causes consumers to mark the current
value as dirty.

There are currently no restricted contexts for `onChange`. It's
conceivable we may want to restrict it.

(cherry picked from commit 9bc421ff01ad1e54db3b7dd75265e8994af89a23)